+++ client/apps/post_grades/components/post_grades/_post_grades.jsx
@@ -7,8 +7,8 @@ import moment from 'moment';
 import { listAssignments } from 'atomic-canvas/libs/constants/assignments';
 import { listCourseSections } from 'atomic-canvas/libs/constants/sections';
 import canvasRequest from 'atomic-canvas/libs/action';
-import { createStudentInfo } from '../actions/submissions';
-import { createSectionInfo, updateSectionMetadata } from '../actions/sections_info';
+import { createStudentInfo } from '../../actions/submissions';
+import { createSectionInfo, updateSectionMetadata } from '../../actions/sections_info';
 
 const select = state => ({
   lmsCourseId: state.settings.lms_course_id,
